*CHANGE4*

+++ src/periodtype.py
@@ -1,18 +1,18 @@
 #data = {
-# "region":
-#{
-#   "name": "Africa",
-#  "avgAge": 19.7,
-##"avgDailyIncomePopulation": 0.71
-#},
-#"periodType": "weeks",
-#"timeToElapse": 7,
-#"reportedCases": 784,
-#"population": 66622705,
-#"totalHospitalBeds": 1380614
+ # "region":
+  #{
+   #   "name": "Africa",
+    #  "avgAge": 19.7,
+     ##"avgDailyIncomePopulation": 0.71
+  #},
+  #"periodType": "weeks",
+  #"timeToElapse": 7,
+  #"reportedCases": 784,
+  #"population": 66622705,
+  #"totalHospitalBeds": 1380614
 #}
 
-from math import  trunc
+from math import trunc
 
 
 def days(data):
@@ -24,11 +24,11 @@ def days(data):
               'currentlyInfected': data['reportedCases']* 50,
               }
         impact['infectionsByRequestedTime'] = impact['currentlyInfected'] * pow(2, int(data['timeToElapse']/3))
-        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int(data['timeToElapse']/ 3))   
-        impact['severeCasesByRequestedTime'] = 0.15 * impact['infectionsByRequestedTime']
-        severeImpact['severeCasesByRequestedTime'] = 0.15 * severeImpact['infectionsByRequestedTime']
+        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int((data['timeToElapse']/ 3)))   
+        impact['severeCasesByRequestedTime'] = int(0.15 * impact['infectionsByRequestedTime'])
+        severeImpact['severeCasesByRequestedTime'] = int(0.15 * severeImpact['infectionsByRequestedTime'])
         impact['hospitalBedsByRequestedTime'] = trunc(availablebeds - impact['severeCasesByRequestedTime'])
-        severeImpact['hospitalBedsByRequestedTime'] =trunc(availablebeds - severeImpact['severeCasesByRequestedTime'])
+        severeImpact['hospitalBedsByRequestedTime'] = trunc(availablebeds - severeImpact['severeCasesByRequestedTime'])
         impact['casesForICUByRequestedTime'] = int(0.05 * impact['infectionsByRequestedTime'])
         severeImpact['casesForICUByRequestedTime'] = int(0.05 * severeImpact['infectionsByRequestedTime'])
         impact['casesForVentilatorsByRequestedTime'] = int(0.02 * impact['infectionsByRequestedTime'])
@@ -51,11 +51,11 @@ def weeks(data):
               'currentlyInfected': data['reportedCases']* 50,
               }
         impact['infectionsByRequestedTime'] = impact['currentlyInfected'] * pow(2, int((data['timeToElapse']*7)/3))
-        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int((data['timeToElapse']*7)/ 3)
-        impact['severeCasesByRequestedTime'] = 0.15 * impact['infectionsByRequestedTime']
-        severeImpact['severeCasesByRequestedTime'] = 0.15 * severeImpact['infectionsByRequestedTime']
-        impact['hospitalBedsByRequestedTime'] =trunc(availablebeds - impact['severeCasesByRequestedTime'])
-        severeImpact['hospitalBedsByRequestedTime'] =trunc(availablebeds - severeImpact['severeCasesByRequestedTime'])
+        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int((data['timeToElapse']*7)/ 3))
+        impact['severeCasesByRequestedTime'] = int(0.15 * impact['infectionsByRequestedTime'])
+        severeImpact['severeCasesByRequestedTime'] = int(0.15 * severeImpact['infectionsByRequestedTime'])
+        impact['hospitalBedsByRequestedTime'] = trunc(availablebeds - impact['severeCasesByRequestedTime'])
+        severeImpact['hospitalBedsByRequestedTime'] = trunc(availablebeds - severeImpact['severeCasesByRequestedTime'])
         impact['casesForICUByRequestedTime'] = int(0.05 * impact['infectionsByRequestedTime'])
         severeImpact['casesForICUByRequestedTime'] = int(0.05 * severeImpact['infectionsByRequestedTime'])
         impact['casesForVentilatorsByRequestedTime'] = int(0.02 * impact['infectionsByRequestedTime'])
@@ -77,10 +77,10 @@ def months(data):
         severeImpact = {
               'currentlyInfected': data['reportedCases']* 50,
               }
-        impact['infectionsByRequestedTime'] = impact['currentlyInfected'] * pow(2, int(data['timeToElapse']*30)/ 3)
-        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int((data['timeToElapse']*30))/ 3)
-        impact['severeCasesByRequestedTime'] = 0.15 * impact['infectionsByRequestedTime']
-        severeImpact['severeCasesByRequestedTime'] = 0.15 * severeImpact['infectionsByRequestedTime']
+        impact['infectionsByRequestedTime'] = impact['currentlyInfected'] * pow(2, int((data['timeToElapse']*30)/ 3))
+        severeImpact['infectionsByRequestedTime'] = severeImpact['currentlyInfected'] * pow(2, int((data['timeToElapse']*30)/ 3))
+        impact['severeCasesByRequestedTime'] = int(0.15 * impact['infectionsByRequestedTime'])
+        severeImpact['severeCasesByRequestedTime'] = int(0.15 * severeImpact['infectionsByRequestedTime'])
         impact['hospitalBedsByRequestedTime'] = trunc(availablebeds - impact['severeCasesByRequestedTime'])
         severeImpact['hospitalBedsByRequestedTime'] = trunc(availablebeds - severeImpact['severeCasesByRequestedTime'])
         impact['casesForICUByRequestedTime'] = int(0.05 * impact['infectionsByRequestedTime'])
